move swap tests to swap_helpers
